🎨 (feature-api) Moved report call to report router file
